extract styled components
